fix: non-interactive database connection failures
Two problems here:
1) endlessly attempting to reconnect (with no interval between attempts)
   is a nightmare that will use resources and spam logs with no respite
2) attempting to update the database credentials on a connection failure
   is definitely not the right thing to do, as they cannot be requested
   when non-interactive.

Problem 1 is not really an issue in interactive use, because the user
being asked for credentials will slow down the connection attempts.

Problem 2 _could_ conceiveably be a problem in interactive mode (if
there's some sort of transient connection failure), but if this is a
problem, it should perhaps be discussed & addressed separately.

Fixes #1294