feat: implement and use new confg key for trusted hosts
After commit https://github.com/koel/koel/commit/e969549, Koel accepts only the hostname from APP_URL (and all its sub domains) as trusted hosts, as long as HTTPS is not enforced. This breaks access with .env.example, until APP_URL is set, and then allows to define a single trusted host only, which can be a problem instances which one wants to access from localhost, from within LAN via local hostname or IP, and remotely with a public hostname at the same time, or for testing instances.

This commit introduces a new config key "TRUSTED_HOSTS". It is empty by default, which permits access via all hostnames, restoring the pre-v7.0.0 behaviour. When definitng it as comma-separated list of hostnames (and/or IPs), access is restricted to those.

Signed-off-by: MichaIng <micha@dietpi.com>